Rename the braintree specific "spec_helper"
The original name was only suffixed "_helper", using "_spec_helper"
is more correct.
