+++ .rspec
@@ -1,2 +1,2 @@
 --color
---require spec_helper
+--require solidus_braintree_spec_helper
