+++ spec/controllers/solidus_braintree/checkouts_controller_spec.rb
@@ -1,4 +1,4 @@
-require 'solidus_braintree_helper'
+require 'solidus_braintree_spec_helper'
 require 'support/solidus_braintree/order_ready_for_payment'
 
 RSpec.describe SolidusBraintree::CheckoutsController, type: :controller do
