+++ spec/controllers/solidus_braintree/client_tokens_controller_spec.rb
@@ -1,4 +1,4 @@
-require 'solidus_braintree_helper'
+require 'solidus_braintree_spec_helper'
 
 RSpec.describe SolidusBraintree::ClientTokensController do
   routes { SolidusBraintree::Engine.routes }
