+++ spec/helpers/solidus_braintree/braintree_admin_helper_spec.rb
@@ -1,4 +1,4 @@
-require 'solidus_braintree_helper'
+require 'solidus_braintree_spec_helper'
 
 RSpec.describe SolidusBraintree::BraintreeAdminHelper do
   describe '#braintree_transaction_link' do
