+++ spec/models/solidus_braintree/avs_result_spec.rb
@@ -1,4 +1,4 @@
-require 'solidus_braintree_helper'
+require 'solidus_braintree_spec_helper'
 
 RSpec.describe SolidusBraintree::AVSResult do
   describe 'AVS response message' do
