+++ spec/models/solidus_braintree/transaction_address_spec.rb
@@ -1,4 +1,4 @@
-require 'solidus_braintree_helper'
+require 'solidus_braintree_spec_helper'
 
 RSpec.describe SolidusBraintree::TransactionAddress do
   describe "#valid?" do
