+++ spec/models/solidus_braintree/transaction_import_spec.rb
@@ -1,4 +1,4 @@
-require 'solidus_braintree_helper'
+require 'solidus_braintree_spec_helper'
 
 RSpec.describe SolidusBraintree::TransactionImport do
   let(:order) { Spree::Order.new }
