+++ spec/models/spree/store_spec.rb
@@ -1,4 +1,4 @@
-require 'solidus_braintree_helper'
+require 'solidus_braintree_spec_helper'
 
 RSpec.describe Spree::Store do
   describe 'before_create :build_default_configuration' do
