+++ spec/solidus_braintree_spec_helper.rb
