+++ spec/system/frontend/braintree_credit_card_checkout_spec.rb
@@ -1,4 +1,4 @@
-require 'solidus_braintree_helper'
+require 'solidus_braintree_spec_helper'
 
 RSpec.shared_context "with frontend checkout setup" do
   let(:braintree) { new_gateway(active: true) }
