+++ spec/system/frontend/venmo_checkout_spec.rb
@@ -1,6 +1,6 @@
 # frozen_string_literal: true
 
-require 'solidus_braintree_helper'
+require 'solidus_braintree_spec_helper'
 
 RSpec.describe "Checkout", type: :feature, js: true do
   let(:braintree_preferences) { { venmo: true }.merge(preferences) }
